+++ fdtd.hpp
@@ -2,4 +2,4 @@
 
 #include "grid.hpp"
 
-int fdtd(dim_t, dim_t, dim_t, int);
+int fdtd(dim_t, dim_t, dim_t, unsigned long);
